Feat/heatmap report (#76)
* feat: post per day report

* fix: tests

* feat: add generic repository and tests

* fix: query

+++ src/main/java/tech/onlycoders/backend/repository/LanguageRepository.java
@@ -19,5 +19,5 @@ public interface LanguageRepository extends Neo4jRepository<Language, String> {
   void setUserLanguage(String canonicalName, String code);
 
   @Query("MATCH (u:User)-[:SPEAKS]->(:Language{code:$code}) RETURN count(u)")
-  Integer getLanguageUseQuantity(String code);
+  Long getLanguageUseQuantity(String code);
 }
